VoiceCloningApplication/requirements.txt
# Core dependencies from coqui-ai
numpy==1.22.0;python_version<="3.10"
numpy>=1.24.3;python_version>"3.10"
cython>=0.29.30
scipy>=1.11.2
torch>=2.1
torchaudio
soundfile>=0.12.0
librosa>=0.10.0
scikit-learn>=1.3.0
numba==0.55.1;python_version<"3.9"
numba>=0.57.0;python_version>="3.9"
inflect>=5.6.0
tqdm>=4.64.1
anyascii>=0.3.0
pyyaml>=6.0
fsspec>=2023.6.0
aiohttp>=3.8.1
packaging>=23.1
mutagen==1.47.0

# Dependencies for examples
flask>=2.0.1

# Dependencies for inference
pysbd>=0.3.4

# Dependencies for notebooks
umap-learn>=0.5.1
pandas>=1.4,<2.0

# Dependencies for training
matplotlib>=3.7.0

# Coqui stack
trainer>=0.0.36

# Config management
coqpit>=0.0.16

# Chinese G2P dependencies
jieba
pypinyin

# Korean dependencies
hangul_romanize

# Gruut + supported languages
gruut[de,es,fr]==2.2.3

# Additional dependencies for Korean
jamo
nltk
g2pkk>=0.1.1

# Dependencies for Bangla
bangla
bnnumerizer
bnunicodenormalizer

# Dependencies for Tortoise
einops>=0.6.0
transformers>=4.33.0

# Dependencies for Bark
encodec>=0.1.1

# Dependencies for XTTS
unidecode>=1.3.2
num2words
spacy[ja]>=3

TTS

# Additional dependencies for your project
Pillow==9.3.0
sounddevice
wavio
moviepy==1.0.3
pydub
pygame
